+++ public/express.js
@@ -1,13 +1,13 @@
 const app = express();
 
-dotenv.config({ path: './env'})
+dotenv.config({ path: './env'});
 
 const db = mysql.createConnection({
     host: process.env.DATABASE_HOST,
     user: process.env.DATABASE_USER,
     password: process.env.DATABASE_PASSWORD,
     database: process.env.DATABASE
-})
+});
 
 db.connect((error) => {
     if(error) {
@@ -15,4 +15,4 @@ db.connect((error) => {
     } else {
         console.log("MySQL connected!")
     }
-})
+});
