+++ public/style3.css
@@ -3,13 +3,13 @@
     padding: 0;
     box-sizing: border-box;
   }
-  
+
   body, html {
     height: 100%;
     font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
     background-color: #f0f0f0;
   }
-  
+
   .container {
     display: flex;
     justify-content: center;
@@ -17,7 +17,7 @@
     min-height: 100vh;
     padding: 2rem;
   }
-  
+
   .card, .details {
     background-color: #grey;
     color: black;
@@ -26,29 +26,29 @@
     flex-basis: 300px;
     border-radius: 8px;
   }
-  
+
   .image-placeholder {
     width: 100%;
     height: 200px;
     background-color: #fff;
     border-radius: 50%;
     margin-bottom: 1rem;
-  }
-  
+  };
+
   .info-caption {
     font-weight: bold;
     text-align: center;
     margin-top: 1rem;
   }
-  
+
   .details {
     background: none;
   }
-  
+
   .details h1 {
     margin-bottom: 1rem;
   }
-  
+
   .details p {
     margin-bottom: 2rem;
     color: black;
@@ -163,4 +163,3 @@
     text-decoration: none;
     margin: 10px 0;
   }
-
