+++ routes/auth.js
@@ -5,4 +5,7 @@ const router = express.Router();
 router.post('/register', authController.register);
 
 
+
+
+
 module.exports = router;
