+++ routes/pages.js
@@ -9,4 +9,10 @@ router.get('/register', (req, res) => {
     res.render('register');
 });
 
+router.get('/login', (req, res) => {
+    res.render('login');
+});
+
+
+
 module.exports = router;
