--- server/.env
@@ -1,4 +1,0 @@
-DATABASE = login_db
-DATABASE_HOST = localhost
-DATABASE_ROOT = root
-DATABASE_PASSWORD =
