--- server/package.json
@@ -1,13 +1,0 @@
-{
-  "dependencies": {
-    "bcryptjs": "^2.4.3",
-    "cookie-parser": "^1.4.6",
-    "dotenv": "^16.4.5",
-    "express": "^4.18.3",
-    "hbs": "^4.2.0",
-    "jsonwebtoken": "^9.0.2",
-    "mysql": "^2.18.1",
-    "nodemon": "^3.1.0"
-
-  }
-}
