--- sprint 2 frontend presentatie/sign-up.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <link rel="stylesheet" href="sign-up.css">
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>Ronde Balk in Formulier</title>
-</head>
-<body> 
-    
-
-<form target="_blank" method="post">
-    <img class="plaatje" src="img/Logo2.gereedschap.jpeg">
-    <p style="text-align: center;">Share Tools</p>
-    <p style="text-align: center;">Sign-up</p>
-
-    <input type="text" name="username" placeholder="Naam" required>
-    <input type="text" name="birthdate" placeholder="Geboortedatum" required>
-    <input type="email" name="email" placeholder="E-mail" required>
-    <input type="password" name="password" placeholder="Wachtwoord" required>
-    <input type="text" name="residence" placeholder="Woonplaats" required>
-    <input onclick="window.location.href = 'index 2.html'" type="submit" value="Sign-up" href="index 2.html">
-</form>
-
-</body>
-</html>
