+++ views/index.hbs
@@ -5,7 +5,7 @@
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <title>Gereedschap App</title>
 <script src="script.js" defer></script>
-<link rel="stylesheet" href="style2.css">
+<link rel="stylesheet" href="/style2.css">
 </head>
 <body>
   <div class="popup-container" id="reminderPopup">
@@ -16,8 +16,8 @@
   <div>
     <a href="#">Lenen</a>
     <a href="#">Help</a>
-    <a href="login.html">Inloggen</a>
-    <a href="sign-up.html">Aanmelden</a>
+    <a href="./login">Inloggen</a>
+    <a href="./register">Aanmelden</a>
   </div>
 </div>
