+++ views/login.hbs
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <link rel="stylesheet" href="login.css">
+    <link rel="stylesheet" href="/login.css">
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <title>Ronde Balk in Formulier</title>
@@ -10,13 +10,13 @@
 
 
 <form target="_blank" method="post">
-    <img class="plaatje" src="img/Logo2.gereedschap.jpeg">
+    <img class="plaatje" src="/img/Logo2.gereedschap.jpeg">
     
     <p style="text-align: center;">Share Tools</p>
     <p style="text-align: center;">Login</p>
-    <input type="text" name="username" placeholder="Gebruikersnaam" required>
-    <input type="email" name="email" placeholder="E-mailadres" required>
-    <input type="password" name="password" placeholder="Wachtwoord" required>
+    <input type="text" id="name" name="username" placeholder="naam" required>
+    <input type="email" id="email" name="email" placeholder="E-mailadres" required>
+    <input type="password"  id="password" name="password" placeholder="Wachtwoord" required>
     <input type="checkbox" checked="checked" name="remember"> Remember me
     <input onclick="window.location.href = 'index 2.html'" type="submit" value="Sign-up" href="index 2.html">
