+++ views/productinfopagina.html
@@ -29,6 +29,6 @@
             <button type="button">BERICHT</button>
           </div>
         </aside>
-      </main>     
+      </main>
 </body>
 </html>
