+++ views/script.js
